more clean up

+++ source/common/http/conn_pool_base.h
@@ -207,7 +207,7 @@ public:
                               OptRef<Upstream::Host::CreateConnectionData> data);
   ~MultiplexedActiveClientBase() override = default;
   // Caps max streams per connection below 2^31 to prevent overflow.
-  static uint64_t maxStreamsPerConnection(uint64_t max_streams_config);
+  static uint32_t maxStreamsPerConnection(uint32_t max_streams_config);
 
   // ConnPoolImpl::ActiveClient
   bool closingWithIncompleteStream() const override;
